Don't try to delete an already deleted invite

+++ lib/utils/brute_utils.dart
@@ -3,18 +3,23 @@ import 'dart:async';
 import 'package:nyxx/nyxx.dart';
 
 class BruteUtils {
+  static bool isRunning = false;
+
   static Future<void> startBrute(
     IGuildChannel targetChannel,
     RegExp regex, {
     required ITextGuildChannel notifyChannel,
   }) async {
+    isRunning = true;
     // Create the invite
-    IInvite invite =
+    IInvite? invite =
         await targetChannel.createInvite(maxAge: 0, maxUses: 0, unique: true);
     // If it doesn't match, delete it and generate a new one
-    while (!regex.hasMatch(invite.code)) {
+    while (invite == null || !regex.hasMatch(invite.code)) {
       try {
-        await invite.delete();
+        if (invite != null) {
+          await invite.delete();
+        }
         invite = await targetChannel.createInvite(maxAge: 0, maxUses: 0, unique: true);
         print(invite.code);
         await Future.delayed(Duration(seconds: 3));
@@ -26,10 +31,12 @@ class BruteUtils {
                 '```\n'
                 '$st'
                 '```'));
+        invite = null;
       }
     }
     await notifyChannel.sendMessage(
         MessageBuilder.content('Finished brute forcing invite for <#${targetChannel.id}>: ${invite.url}'));
+    isRunning = false;
     return;
   }
 }
